Handle overflow conditions with narrow bandwidth
When given a narrow range of audio frequencies (200 Hz to 4000 Hz specifically),
I was getting a crash due to overflowing the limits of short signed ints.

Since the frequency limits of my transmission from an FM handheld will be about
in this range, it's pretty crucial to be able to narrow the upper bandwidth
limit down from 20 kHz to 4 kHz to produce something I can theoretically
transmit without clipping the image down to the bottom 20% on the receiver's
waterfall.

To fix this, catch the OverflowError exception and set the values to short int
min or max depending on the sign of the value being aggregated into the current
audio sample.