Support optional image rotation
It looks correct on the audacity spectrograph because frequency is
the Y axis and time is the X axis there, but most ham radio
spectrographs (waterfalls) have frequency for the X axis and time
for the Y axis.

Therefore, let's make it a boolean CLI argument defaulting to the
legacy behavior.